Remove now unused .linux. and .windows. suffixes
Combine classes into one file, where still needed
