--- src/System.Device.Gpio/System/Device/Gpio/Drivers/HummingBoardDriver.Windows.cs
@@ -1,10 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace System.Device.Gpio.Drivers
-{
-    public partial class HummingBoardDriver : Windows10Driver
-    {
-    }
-}
