+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/InterruptSysFsDriver.cs
