+++ src/System.Device.Gpio/System/Device/Gpio/LibgpiodDriverEventHandler.cs
