+++ src/System.Device.Gpio/System/Device/I2c/Devices/UnixI2cDevice.cs
