+++ src/System.Device.Gpio/System/Device/I2c/Devices/Windows10I2cDevice.cs
