+++ src/System.Device.Gpio/System/Device/Pwm/Channels/UnixPwmChannel.cs
