+++ src/System.Device.Gpio/System/Device/Pwm/Channels/Windows10PwmChannel.cs
