+++ src/System.Device.Gpio/System/Device/Spi/Devices/UnixSpiDevice.cs
