+++ src/System.Device.Gpio/System/Device/Spi/Devices/Windows10SpiDevice.cs
